bot/requirements.txt
python-telegram-bot==21.10
aiofiles==24.1.0
aiosqlite==0.20.0
async-timeout==5.0.1
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
typing_extensions==4.12.2
